Added data marks, user scaling and started adjustable rectangle for bar chart.
Line are now drawn when graph is created, done in increments of 1 or 2 depending on the highest value. User can now use slider to change size of graph, this saves need for automated scaler system let the user do it themselves. Work has begun on making a rectangle that can be created with variables to make the height adjustable for a bar chart concept.