Added Freetype library. Minor Changes as to camera and rotation system.
Started to add the freetype library, files and links added to get it working. Made minor changes to the rotation so program doesn't jolt when moving for the second time. Added keyboard controls to move camera along x and z axes to allow user to get better view of far away data points.