Added Fonts and Testing if Freetype can load.
Tested if freetype can run and if program can find the libraries. After some error debugging, found that I had built the freetype lib in x64 instead of win32 (program is built in win32). Freetype now loads correctly and can load in a font. Fonts have been added to a folder in the project to load from.