Labels for all 3 Axes (With Memory Leak method)
labels can now work for the 3 axes but the method leaks a lot of memory and cannot be used. Trying to create a vector of quads to make the quads out-with the main render loop, preventing the leaking memory. Currently facing an issue whereby quads are not getting made correctly.