Implemented Colour Picker for user. Need to fix bar chart due to axes flipping.
User can now choose which colour they want the data points to be by using colour picker in the GUI. Found that the bar chart creation was incorrect since data was being used on the z axes instead of the y axes. Need to get bar chart working in Y direction or keep using z axes. Also need to put labels for the x,y and z axes.