Trying to get /MT to build. Still no success.
Been messing about with libraries and such, still no working build with /MT enabled as the code generator. Still amazed that still switching between /MT and /MD causes the program to run fine. Must be an issue with either the Libs themselves (they are dynamically linked) or a small brain moment with the library linking with the program.